Use new helper for dashboards API
